requirements.txt
anyio==4.6.2.post1
bech32==1.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cosmospy==6.0.0
ecdsa==0.19.0
exceptiongroup==1.2.2
h11==0.14.0
hdwallets==0.1.2
httpcore==1.0.6
httpx==0.27.2
idna==3.10
mantrapy @ file:///home/rama/hack/mantrapy
mnemonic==0.21
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-telegram-bot==21.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
